refactor: move config utils to separate module and add xml validation config
* Introduced new XML validation configuration option for max iterations
* Centralized configuration utilities in a dedicated module
* Updated all imports to use the new config utils module location
