+++ src/agent/modelHandlerAnthropic.ts
@@ -29,7 +29,7 @@ import {
 import { ToolState } from './ToolState';
 import { AgentStateRound } from './AgentState';
 import { messageToSkeleton } from './messageUtils';
-import { getConfig } from '../frontend-utils/commonUtils';
+import { getConfig } from '../utils/configUtils';
 
 const K_SLICE = 200;
